aplicativo/requirements.txt
torch
torchvision
Flask
Pillow
gunicorn
gdown
typing-extensions